feat: add enum radio and select widgets for schema forms
- Implemented `EnumRadio` component for enum fields with 3 or fewer options, displaying descriptions for each option.
- Implemented `EnumSelect` component for enum fields with more than 3 options, utilizing native browser tooltips for descriptions.
- Added `IntegerField` and `NumberField` components for handling integer and decimal inputs, respectively, with validation for min/max values.
- Introduced `GcodePreview` service to manage G-code parsing and layer visibility in the UI.
- Created `Logger` service for scoped logging throughout the application.
- Refactored `SceneCommand` and `SceneEngine` services to improve scene operation handling and logging.
- Updated `SceneHistory` service to utilize the new `SceneEngine`.
- Modified `Slicer` service to integrate with the updated `SceneEngine`.
- Enhanced `TooltipDirective` and created a new `Tooltip` component for improved tooltip functionality.
- Updated environment configuration to switch slicing backend from 'wasm' to 'server'.
