+++ ui/src/app/app.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { NotificationCenterComponent } from './components/notification-center/notification-center.component';
+import { NotificationCenter } from './components/notification-center/notification-center';
 
 @Component({
   selector: 'nexus-root',
   standalone: true,
-  imports: [RouterOutlet, NotificationCenterComponent],
+  imports: [RouterOutlet, NotificationCenter],
   templateUrl: './app.html',
   styleUrl: './app.scss',
 })
