+++ ui/src/app/components/3d-view-toolbar/3d-view-toolbar.html
@@ -22,7 +22,7 @@
     [tooltip]="viewMode() === 'gcode' ? 'Switch to model view' : 'Switch to G-code preview'"
     (click)="toggleViewMode()"
   >
-    <nexus-icon [name]="viewMode() === 'gcode' ? 'cube' : 'layers'"></nexus-icon>
+    <nexus-icon [name]="viewMode() === 'gcode' ? 'cube' : 'multiple-pages'"></nexus-icon>
   </button>
 </div>
