+++ ui/src/app/components/settings-panel/settings-panel.ts
@@ -1,6 +1,6 @@
 import { Component, inject } from '@angular/core';
 import globalSettingsSchema from '../../../schemas/slicer-engine-global-settings-v1.json';
-import { FieldChangeEvent, SchemaFormComponent } from '../../schema-form/schema-form.component';
+import { FieldChangeEvent, SchemaForm } from '../../schema-form/schema-form';
 import { Slicer } from '../../services/slicer';
 
 // Extract the SlicingParams sub-schema so the form renders all slicer settings.
@@ -15,11 +15,11 @@ const SLICING_PARAMS_SCHEMA = {
 @Component({
   selector: 'nexus-settings-panel',
   standalone: true,
-  imports: [SchemaFormComponent],
+  imports: [SchemaForm],
   templateUrl: './settings-panel.component.html',
   styleUrl: './settings-panel.component.scss',
 })
-export class SettingsPanelComponent {
+export class SettingsPanel {
   private readonly slicer = inject(Slicer);
 
   readonly settings = this.slicer.settings;
