+++ ui/src/app/components/status-panel/status-panel.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, Component, effect, inject, ViewChild } from '@angular/core';
+import { Component, ElementRef, afterNextRender, effect, inject, viewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { Slicer } from '../../services/slicer';
 
@@ -9,10 +9,10 @@ import { Slicer } from '../../services/slicer';
   templateUrl: './status-panel.component.html',
   styleUrl: './status-panel.component.scss',
 })
-export class StatusPanelComponent implements AfterViewInit {
+export class StatusPanel {
   private readonly slicer = inject(Slicer);
 
-  @ViewChild('logContainer') logContainer: any;
+  private readonly logContainer = viewChild<ElementRef<HTMLElement>>('logContainer');
 
   readonly status = this.slicer.status;
   readonly outputLog = this.slicer.outputLog;
@@ -23,27 +23,23 @@ export class StatusPanelComponent implements AfterViewInit {
   constructor() {
     // Auto-scroll when output log changes
     effect(() => {
-      if (this.autoScroll && this.logContainer) {
-        // Use setTimeout to ensure DOM is updated
-        setTimeout(() => {
-          this.scrollToBottom();
-        }, 0);
-      }
-      // Access outputLog to create dependency
       this.outputLog();
+      if (this.autoScroll) {
+        this.scrollToBottom();
+      }
+    });
+
+    afterNextRender(() => {
+      if (this.autoScroll) {
+        this.scrollToBottom();
+      }
     });
   }
 
-  ngAfterViewInit(): void {
-    // Initial scroll to bottom
-    if (this.autoScroll) {
-      this.scrollToBottom();
-    }
-  }
-
   private scrollToBottom(): void {
-    if (this.logContainer) {
-      this.logContainer.nativeElement.scrollTop = this.logContainer.nativeElement.scrollHeight;
+    const el = this.logContainer()?.nativeElement;
+    if (el) {
+      el.scrollTop = el.scrollHeight;
     }
   }
