+++ ui/src/environments/environment.ts
@@ -15,5 +15,5 @@ export const environment: Environment = {
   production: false,
   apiUrl: `${httpProtocol}//${host}:${BACKEND_PORT}/api`,
   wsUrl: `${wsProtocol}//${host}:${BACKEND_PORT}/ws`,
-  sliceBackend: 'wasm',
+  sliceBackend: 'server',
 };
